🎨 代码优化

+++ src/main/java/com/rymcu/forest/lucene/service/impl/UserLuceneServiceImpl.java
@@ -161,7 +161,7 @@ public class UserLuceneServiceImpl implements UserLuceneService {
         }
         resList.add(
             UserLucene.builder()
-                .idUser(Integer.valueOf(hitDoc.get("id")))
+                .idUser(Long.valueOf(hitDoc.get("id")))
                 .nickname(titleValue.toString())
                 .signature(baikeValue.toString())
                 .score(String.valueOf(score))
@@ -182,7 +182,7 @@ public class UserLuceneServiceImpl implements UserLuceneService {
   }
 
   @Override
-  public List<UserDTO> getUsersByIds(Integer[] ids) {
+  public List<UserDTO> getUsersByIds(Long[] ids) {
     return userLuceneMapper.getUsersByIds(ids);
   }
 }
